create controller
